Create folders to organise components by view
