--- public/components/Resources.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const Resources = () => <h1>Resources</h1>;
-
-export default Resources;
