+++ public/components/Resources/Resources.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import Navbar from '../Navbar';
+
+
+const Resources = () => (
+  <React.Fragment>
+    <Navbar />
+    <h1>Resources</h1>
+  </React.Fragment>
+);
+
+export default Resources;
