+++ public/style.css
@@ -101,8 +101,9 @@ h1 {
   flex-direction: column;
   background-image: url("./shape.svg");
   background-repeat: no-repeat;
-  background-position: -20% -10%;
   background-size: 100%; }
+  .home .navbar {
+    background-color: transparent; }
   .home .home_title {
     width: 100%; }
     .home .home_title h1 {
@@ -130,11 +131,16 @@ h1 {
     -webkit-box-align: center;
     -ms-flex-align: center;
     align-items: center;
-    justify-content: space-around; }
+    justify-content: space-around;
+    margin-bottom: 20%; }
     .home .third_section .home_button {
       border: 2px solid #8145A5;
       width: 160px;
       margin: 2%; }
+  .home footer {
+    width: 100%;
+    height: 10%;
+    background-color: #fc7844; }
 
 button {
   background-color: #FFF6F1;
