+++ public/style.scss
@@ -137,9 +137,12 @@ h1 {
       -ms-flex-direction: column; 
           flex-direction: column; 
     background-image : url("./shape.svg"); 
-    background-repeat: no-repeat; 
-    background-position: -20% -10%; 
+    background-repeat: no-repeat;  
     background-size: 100%; 
+
+    & .navbar {
+        background-color: transparent; 
+    }
 
     & .home_title {
         width: 100%; 
@@ -184,12 +187,19 @@ h1 {
             -ms-flex-align: center;
                 align-items: center; 
         justify-content: space-around;
+        margin-bottom: 20%;  
         
         & .home_button {
             border: 2px solid $purple; 
             width: 160px; 
             margin: 2%; 
         }
+    }
+
+    & footer {
+        width: 100%; 
+        height: 10%;
+        background-color: $orange; 
     }
 }
